pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trading_system"
version = "0.1.0"
description = "A systematic hedge fund architecture project"
authors = [
    { name = "Alex B" }
]
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "pyarrow",
    "numpy",
    "cvxpy",
    "scipy",
    "PyYAML",
    "python-dotenv",
    "scikit-learn",
    "alpaca-py",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "pre-commit",
]

[tool.ruff]
line-length = 79
target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (E) and pycodestyle (F) codes by default.
select = ["E", "F", "I", "N", "PL"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true
